Handle the DEX format's modified UTF-8
Dalvik, why are you doing this to me? I thought we were friends! :(

I was pleasantly surprised to find I could do this as an error handler
instead of a complete codec implementation, though. Yay, Python!

From the docs:
  https://source.android.com/devices/tech/dalvik/dex-format.html

As a concession to easier legacy support, the .dex format encodes its
string data in a de facto standard modified UTF-8 form, hereafter
referred to as MUTF-8. This form is identical to standard UTF-8,
except:

 * Only the one-, two-, and three-byte encodings are used.
 * Code points in the range U+10000 … U+10ffff are encoded as a
   surrogate pair, each of which is represented as a three-byte
   encoded value.
 * The code point U+0000 is encoded in two-byte form.
 * A plain null byte (value 0) indicates the end of a string, as is
   the standard C language interpretation.

This commit's soundtrack: Roxette - Opportunity Nox